fixes - description
Problem with duplicates; the registrar has the ability to execute x orders for 1 registration number, additionally registrars are distinguished by the registration number - there is no situation of overwriting registrars for 1 registration number prefix. Added "city" field instead of contact person, added a field to search for a registrar by city (also searches by address field for better precision). It is possible to link the registration number to the city.